requirements.txt
fastapi
uvicorn
python-jose
pika
redis
pydantic 